requirements.txt
# Legal Document AI Summarizer - Requirements

# Legal Document AI Summarizer - Requirements (Updated for Latest LangChain)

# Core AI and ML libraries
langchain>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
google-generativeai>=0.3.0
faiss-cpu>=1.7.4
numpy>=1.24.0
torch
# Document processing
unstructured>=0.10.0
pypdf>=3.17.0
python-docx>=0.8.11
docx2txt>=0.8

# Image processing for PDF handling
pdf2image>=1.0.0
Pillow>=10.0.0

# Environment and utilities
python-dotenv>=1.0.0

# Additional dependencies
python-dateutil>=2.8.2

# Optional: GPU support (uncomment if you have CUDA)
# faiss-gpu>=1.7.4